Implement message Queue in Postgres
When a document is uploaded, a message is inserted in the queue. Then a poller reads those messages every 10' and triggers auto-training for the projects that have enable it.

For context: Didn't want to add a message queue at this stage like AMQ or RabbitMQ. Postgres is good enough until having 1000s files per second

- Add DB table, JPA entity, and Repository for queue_message
- Add simple message producer
- Add simple message consumer that loads multiple messages in batches.